- fix: Direct ERC-223 swaps through the Pool support for deadline/slippage tolerance - tests for this fix - PeripheryValidation.sol moved to Libraries - _blockTimestamp in all cases return uint32

+++ contracts/libraries/BlockTimestamp.sol
@@ -6,7 +6,7 @@ pragma solidity =0.7.6;
 abstract contract BlockTimestamp {
     /// @dev Method that exists purely to be overridden for tests
     /// @return The current block timestamp
-    function _blockTimestamp() internal view virtual returns (uint256) {
-        return block.timestamp;
+    function _blockTimestamp() internal view virtual returns (uint32) {
+        return uint32(block.timestamp);
     }
 }
