+++ contracts/test/MockTimeNonfungiblePositionManager.sol
@@ -13,8 +13,8 @@ contract MockTimeNonfungiblePositionManager is DexaransNonfungiblePositionManage
 //        address _tokenDescriptor
     ) DexaransNonfungiblePositionManager(_factory, _WETH9) {}
 
-    function _blockTimestamp() internal view override returns (uint256) {
-        return time;
+    function _blockTimestamp() internal view override returns (uint32) {
+        return uint32(time);
     }
 
     function setTime(uint256 _time) external {
